+++ changelog.md
@@ -6,6 +6,7 @@ Features
 * Offer filename completions on more special commands, such as `\edit`.
 * Allow styling of status, timing, and warnings text.
 * Set up customization of prompt/continuation colors in `~/.myclirc`.
+* Allow customization of the toolbar with prompt format strings.
 
 
 Bug Fixes